Renamed statistics plot functions and added functionality
