+++ mprofile_20201203123740.dat
@@ -1,0 +1,23 @@
+FUNC pySIMDEUM.user.compute_presence 117.6406 1606995462.1052 120.2109 1606995462.1232 1
+FUNC pySIMDEUM.user.compute_presence 120.2109 1606995462.1232 121.7227 1606995462.1392 1
+FUNC pySIMDEUM.house.simulate 121.7227 1606995462.1392 123.6523 1606995462.2042 1
+FUNC __main__.simulate_house 117.5352 1606995462.0947 123.6523 1606995462.2042 0
+CMDLINE C:\Users\hillebr\Miniconda3\envs\pySimdeum\python.exe water_use_tests.py
+MEM 0.636719 1606995460.5380
+MEM 14.792969 1606995460.6389
+MEM 23.148438 1606995460.7389
+MEM 32.972656 1606995460.8397
+MEM 41.558594 1606995460.9403
+MEM 50.289062 1606995461.0403
+MEM 55.761719 1606995461.1408
+MEM 65.066406 1606995461.2413
+MEM 69.765625 1606995461.3418
+MEM 82.605469 1606995461.4419
+MEM 86.144531 1606995461.5419
+MEM 93.027344 1606995461.6426
+MEM 101.367188 1606995461.7431
+MEM 106.546875 1606995461.8431
+MEM 109.445312 1606995461.9432
+MEM 123.933594 1606995462.0432
+MEM 123.882812 1606995462.1437
+MEM 123.660156 1606995462.2437
